+++ .eslintrc.json
@@ -19,7 +19,9 @@
     "react/react-in-jsx-scope": "off",
     "import/no-unresolved": "off",
     "no-shadow": "off",
-    "no-param-reassign": 0
+    "no-param-reassign": 0,
+    "react/prop-types": 0,
+    "react/destructuring-assignment": 0
   },
   "ignorePatterns": ["dist/", "build/"]
 }
