+++ src/pages/MyProfile.js
@@ -1,5 +1,3 @@
-/* eslint-disable max-len */
-
 import { useSelector } from 'react-redux';
 
 const MyProfile = () => {
